+++ store/rootmulti/import.go
@@ -10,8 +10,8 @@ import (
 	"github.com/crypto-org-chain/cronos-store/memiavl"
 
 	cosmoserrors "cosmossdk.io/errors"
-	"cosmossdk.io/store/snapshots/types"
 
+	"github.com/cosmos/cosmos-sdk/store/v2/snapshots/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 )