+++ store/rootmulti/objstore.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/crypto-org-chain/cronos-store/memiavl"
 
-	"cosmossdk.io/store/transient"
-	"cosmossdk.io/store/types"
+	"github.com/cosmos/cosmos-sdk/store/v2/transient"
+	"github.com/cosmos/cosmos-sdk/store/v2/types"
 )
 
 // GetObjKVStore Implements interface MultiStore